move setup config
